# docs/swagger.yaml
basePath: /
definitions:
  github_com_skiix-backend_internal_domain.AddMembersRequest:
    properties:
      user_ids:
        items:
          type: string
        maxItems: 50
        minItems: 1
        type: array
    required:
    - user_ids
    type: object
  github_com_skiix-backend_internal_domain.Circle:
    properties:
      avatar_url:
        example: https://cdn.skiix.com/avatar.jpg
        type: string
      created_at:
        example: "2026-04-10T12:00:00Z"
        type: string
      description:
        example: My inner circle
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      member_count:
        example: 1
        type: integer
      name:
        example: Close Friends
        type: string
      owner_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      updated_at:
        example: "2026-04-10T12:00:00Z"
        type: string
    type: object
  github_com_skiix-backend_internal_domain.CircleMember:
    properties:
      circle_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      joined_at:
        example: "2026-04-10T12:00:00Z"
        type: string
      role:
        example: owner
        type: string
      user_id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
    type: object
...
  github_com_skiix-backend_internal_domain.CreateCircleRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  github_com_skiix-backend_internal_domain.UpdateCircleRequest:
    properties:
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 50
        minLength: 1
        type: string
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Skiix authentication API with JWT and OAuth2 support
  title: Skiix API
  version: "1.0"
paths:
...
  /circles:
    get:
      description: Returns a list of all circles the current user is a member of.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_skiix-backend_internal_domain.Circle'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user circles
      tags:
      - circles
    post:
      consumes:
      - application/json
      description: Creates a new private circle.
      parameters:
      - description: Circle info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_skiix-backend_internal_domain.CreateCircleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_skiix-backend_internal_domain.Circle'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new circle
      tags:
      - circles
  /circles/{id}:
    delete:
      description: Deletes a circle (owner only).
      parameters:
      - description: Circle ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a circle
      tags:
      - circles
    put:
      consumes:
      - application/json
      description: Updates the name/description of a circle (owner only).
      parameters:
      - description: Circle ID
        in: path
        name: id
        required: true
        type: string
      - description: Circle info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_skiix-backend_internal_domain.UpdateCircleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update an existing circle
      tags:
      - circles
  /circles/{id}/members:
    get:
      description: Returns paginated members of a circle (must be a member).
      parameters:
      - description: Circle ID
        in: path
        name: id
        required: true
        type: string
      - description: Pagination limit
        in: query
        name: limit
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_skiix-backend_internal_domain.CircleMember'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get members of a circle
      tags:
      - circles
    post:
      consumes:
      - application/json
      description: Adds multiple users to a circle (owner only).
      parameters:
      - description: Circle ID
        in: path
        name: id
        required: true
        type: string
      - description: Array of user IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_skiix-backend_internal_domain.AddMembersRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add members to a circle
      tags:
      - circles
  /circles/{id}/members/{user_id}:
    delete:
      description: Owner can remove any member, member can remove themselves.
      parameters:
      - description: Circle ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID to remove
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove a member from a circle
      tags:
      - circles
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
    x-tokenname: Authorization
swagger: "2.0"